Untested splitFunction()
I am just finished wiht an optimal design for the splitFunction() but it
is untested.
I need to put this baby in use.
Probalby i will have to create a function that displays the
splitFunctionsArray in a nice format, and to write all this in stone for
future and extended use.